requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
anthropic>=0.40.0
click>=8.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
httpx>=0.27.0
